refactor(polymer): stage 1 — split polymer.py into a package
Convert meeko/polymer.py (3597 LOC, 7 classes in one file) into the
meeko/polymer/ package, one module per concern:

  errors.py       PolymerCreationError                          (37 LOC)
  utils.py        find_graph_paths, find_inter_mols_bonds*,    (589 LOC)
                  mapping_by_mcs, _snap_to_int,
                  divide_int_gracefully, rectify_charges,
                  get_updated_positions, update_H_positions,
                  _delete_residues, handle_parsing_situations
  padder.py       ResiduePadder, NoAtomMapWarning + helpers    (364 LOC)
                  (get_molAtomMapNumbers, etc.)
  templates.py    ResidueTemplate, ResidueChemTemplates        (273 LOC)
  monomer.py      Monomer                                       (295 LOC)
  rotamers.py     residues_rotamers data + the rotamer routine (116 LOC)
  polymer.py      Polymer class (still ~1700 LOC, untouched   (1729 LOC)
                  pending stages 2-3 of the plan)
  __init__.py     backward-compat re-exports                    (74 LOC)

No logic changes. External imports
``from meeko.polymer import Polymer, Monomer, ResiduePadder,
ResidueTemplate, ResidueChemTemplates, PolymerCreationError,
add_rotamers_to_polymer_molsetups`` continue to work unchanged. All 92
tests pass (4 skipped).

This is preparatory: stage 2 will dedupe the three _pdb_to_residue_mols
/ _pqr_to_residue_mols / _prody_to_residue_mols static methods on
Polymer, and stage 3 will pull network template-fetching out of
Polymer.__init__.
