+++ meeko/polymer/rotamers.py
@@ -1,0 +1,116 @@
+"""Rotamer support: residue→sidechain-dihedral table and the routine that
+applies a list of rotamer states to a polymer's monomer molsetups.
+"""
+
+import logging
+
+import numpy as np
+
+logger = logging.getLogger(__name__)
+
+
+residues_rotamers = {
+    "SER": [("C", "CA", "CB", "OG")],
+    "THR": [("C", "CA", "CB", "CG2")],
+    "CYS": [("C", "CA", "CB", "SG")],
+    "VAL": [("C", "CA", "CB", "CG1")],
+    "HIS": [("C", "CA", "CB", "CG"), ("CA", "CB", "CG", "CD2")],
+    "ASN": [("C", "CA", "CB", "CG"), ("CA", "CB", "CG", "ND2")],
+    "ASP": [("C", "CA", "CB", "CG"), ("CA", "CB", "CG", "OD1")],
+    "ILE": [("C", "CA", "CB", "CG2"), ("CA", "CB", "CG2", "CD1")],
+    "LEU": [("C", "CA", "CB", "CG"), ("CA", "CB", "CG", "CD1")],
+    "PHE": [("C", "CA", "CB", "CG"), ("CA", "CB", "CG", "CD2")],
+    "TYR": [("C", "CA", "CB", "CG"), ("CA", "CB", "CG", "CD2")],
+    "TRP": [("C", "CA", "CB", "CG"), ("CA", "CB", "CG", "CD2")],
+    "GLU": [
+        ("C", "CA", "CB", "CG"),
+        ("CA", "CB", "CG", "CD"),
+        ("CB", "CG", "CD", "OE1"),
+    ],
+    "GLN": [
+        ("C", "CA", "CB", "CG"),
+        ("CA", "CB", "CG", "CD"),
+        ("CB", "CG", "CD", "OE1"),
+    ],
+    "MET": [
+        ("C", "CA", "CB", "CG"),
+        ("CA", "CB", "CG", "SD"),
+        ("CB", "CG", "SD", "CE"),
+    ],
+    "ARG": [
+        ("C", "CA", "CB", "CG"),
+        ("CA", "CB", "CG", "CD"),
+        ("CB", "CG", "CD", "NE"),
+        ("CG", "CD", "NE", "CZ"),
+    ],
+    "LYS": [
+        ("C", "CA", "CB", "CG"),
+        ("CA", "CB", "CG", "CD"),
+        ("CB", "CG", "CD", "CE"),
+        ("CG", "CD", "CE", "NZ"),
+    ],
+}
+
+
+def add_rotamers_to_polymer_molsetups(rotamer_states_list, polymer):
+    """Apply a list of rotamer-state dicts to each Monomer.molsetup in
+    ``polymer``. Returns a list (one entry per state) of dicts mapping
+    residue-id → index into ``molsetup.rotamers``.
+    """
+    rotamer_res_disambiguate = {}
+    for (
+        primary_res,
+        specific_res_list,
+    ) in polymer.residue_chem_templates.ambiguous.items():
+        for specific_res in specific_res_list:
+            rotamer_res_disambiguate[specific_res] = primary_res
+
+    no_resname_to_resname = {}
+    for res_with_resname in polymer.monomers:
+        chain, resname, resnum = res_with_resname.split(":")
+        no_resname_key = f"{chain}:{resnum}"
+        if no_resname_key in no_resname_to_resname:
+            errmsg = "both %s and %s would be keyed by %s" % (
+                res_with_resname,
+                no_resname_to_resname[no_resname_key],
+                no_resname_key,
+            )
+            raise RuntimeError(errmsg)
+        no_resname_to_resname[no_resname_key] = res_with_resname
+
+    state_indices_list = []
+    for state_index, state_dict in enumerate(rotamer_states_list):
+        logger.info(f"adding rotamer state {state_index + 1}")
+        state_indices = {}
+        for res_no_resname, angles in state_dict.items():
+            res_with_resname = no_resname_to_resname[res_no_resname]
+            if polymer.monomers[res_with_resname].molsetup is None:
+                raise RuntimeError(
+                    "no molsetup for %s, can't add rotamers" % (res_with_resname)
+                )
+            molsetup = polymer.monomers[res_with_resname].molsetup
+            name_to_molsetup_idx = {}
+            for atom in molsetup.atoms:
+                atom_name = atom.pdbinfo.name
+                name_to_molsetup_idx[atom_name] = atom.index
+
+            resname = res_with_resname.split(":")[1]
+            resname = rotamer_res_disambiguate.get(resname, resname)
+
+            atom_names = residues_rotamers[resname]
+            if len(atom_names) != len(angles):
+                raise RuntimeError(
+                    f"expected {len(atom_names)} angles for {resname}, got {len(angles)}"
+                )
+
+            atom_idxs = []
+            for names in atom_names:
+                tmp = [name_to_molsetup_idx[name] for name in names]
+                atom_idxs.append(tmp)
+
+            state_indices[res_with_resname] = len(molsetup.rotamers)
+            molsetup.add_rotamer(atom_idxs, np.radians(angles))
+
+        state_indices_list.append(state_indices)
+
+    return state_indices_list
